migrate test_chroot to pytest
